LayoutTests/animations/trigger-computed-style-expected.txt

CONSOLE MESSAGE: line 24: TypeError: null is not an object (evaluating 'triggerStyle[i]')
Test the computed style of the -webkit-animation-trigger property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Auto : auto
FAIL triggerStyle.length should be 1. Threw exception TypeError: null is not an object (evaluating 'triggerStyle.length')
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
